backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
distro==1.9.0
Flask==3.0.2
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
Flask-Session==0.6.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mock==5.1.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
openai==1.14.2
packaging==23.2
pluggy==1.4.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
pyotp==2.9.0
pypng==0.20220715.0
pytest==8.0.2
python-dotenv==1.0.1
qrcode==7.4.2
requests==2.31.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
